mcp_kali_full_project/requirements.txt
rich
typer[all]
requests
pyyaml
